rendering sprites

+++ src/Resources/Texture.h
@@ -11,8 +11,8 @@ namespace shen
         void SetPath(const std::string& path) { _path = path; }
         const std::string& GetPath() const { return _path; }
 
-        void SetSize(const glm::ivec2 size) { _size = size; }
-        glm::ivec2 GetSize() const { return _size; }
+        void SetSize(const glm::vec2 size) { _size = size; }
+        glm::vec2 GetSize() const { return _size; }
 
         void SetId(unsigned int id) { _id = id; }
         unsigned int GetId() const { return _id; }
@@ -20,6 +20,6 @@ namespace shen
     private:
         std::string _path;
         unsigned int _id = 0;
-        glm::ivec2 _size;
+        glm::vec2 _size;
     };
 }
